# apps/drive/public/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Xenode Internal API",
    "description": "🔒 **INTERNAL USE ONLY** — This API reference is for internal development and integration purposes. All endpoints require authentication via session cookie or API key bearer token.\n\nXenode is a zero-knowledge encrypted cloud storage platform. The backend never sees raw file contents or unencrypted keys. All E2EE operations (encrypt, decrypt, key wrapping) are performed client-side.",
    "version": "0.1.0",
    "contact": {
      "name": "Xenode Engineering"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current environment"
    }
  ],
...
  "security": [
    { "bearerAuth": [] },
    { "cookieAuth": [] }
  ],
  "paths": {
...
    "/api/objects": {
      "get": {
        "tags": ["Objects"],
        "summary": "List files and folders (paginated)",
        "description": "Returns files and folders with full metadata, signed thumbnail/optimized URLs. Supports pagination, filtering by bucket/folder, content type, and media category.",
        "parameters": [
          { "name": "bucketId", "in": "query", "schema": { "type": "string" } },
          { "name": "folder", "in": "query", "schema": { "type": "string" }, "description": "Direct children of one folder: a folder id or 'root'. Omit to list the whole Space." },
          { "name": "contentType", "in": "query", "schema": { "type": "string" }, "description": "Filter by content type prefix (e.g. 'image')" },
          { "name": "mediaCategory", "in": "query", "schema": { "type": "string" }, "description": "Filter by media category" },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of objects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "objects": { "type": "array", "items": { "$ref": "#/components/schemas/StorageObject" } }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/objects/{id}": {
      "get": {
        "tags": ["Objects"],
        "summary": "Get file download URL",
        "description": "Returns pre-signed download URL (or chunk URLs for chunked files) with encryption metadata.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Download details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": { "type": "string", "format": "uri" },
                    "chunkUrls": { "type": "array", "items": { "type": "string", "format": "uri" } },
                    "isEncrypted": { "type": "boolean" },
                    "encryptedDEK": { "type": "string" },
                    "iv": { "type": "string" },
                    "encryptedName": { "type": "string" },
                    "contentType": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": { "description": "Object not found" }
        }
      },
      "patch": {
        "tags": ["Objects"],
        "summary": "Update object metadata (tags, position)",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tags": { "type": "array", "items": { "type": "string" }, "example": ["work", "important"] },
                  "position": { "type": "integer", "example": 1 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Object updated" },
          "404": { "description": "Not found" }
        }
      },
      "delete": {
        "tags": ["Objects"],
        "summary": "Delete a file or folder",
        "description": "Moves the object into the Bin. Physical deletion and quota release occur through the permanent-deletion lifecycle.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Object deleted" },
          "404": { "description": "Not found" }
        }
      }
    },
...
    "/api/objects/folder": {
      "post": {
        "tags": ["Objects"],
        "summary": "Create a folder",
        "description": "Requires Space manage permission. Folders are metadata records with no blob and the opaque key <spaceRoot><folderId>/. The client chooses the folder id so the encrypted name can be bound to it. Plaintext names are not accepted.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["bucketId", "folderId", "encryptedDisplayName"],
                "properties": {
                  "bucketId": { "type": "string" },
                  "folderId": { "type": "string", "pattern": "^[a-f0-9]{24}$", "description": "New folder id chosen by the client (12 random bytes, lowercase hex)" },
                  "encryptedDisplayName": { "type": "string", "minLength": 16, "maxLength": 4096, "description": "Folder name sealed with the Space metadata key, bound to folderId (xenode-file/1 metadata, purpose name)" },
                  "parentFolderId": { "type": ["string", "null"], "description": "Live parent folder in the same Space; null or omitted for the Space root" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Folder created; body { folder }" },
          "400": { "description": "Missing or malformed folder id, missing encrypted name, or nesting deeper than 32 folders" },
          "409": { "description": "folder_id_conflict: the folder id is already in use" },
          "403": { "description": "Space manage permission required" },
          "404": { "description": "Bucket or parent folder not found in this Space" },
          "409": { "description": "Space inactive or storage usage not initialized" }
        }
      },
      "delete": {
        "tags": ["Objects"],
        "summary": "Move a folder to the Bin",
        "description": "Requires Space delete permission. The folder, its live descendants and their sidecars share one deletedAt so restore and purge act on exactly that batch.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "required": ["folderId"], "properties": { "folderId": { "type": "string" } }
        } } } },
        "responses": {
          "200": { "description": "Folder subtree binned" },
          "404": { "description": "Live folder not found in this Space" }
        }
      }
    },
...
    "/api/objects/thumbnail": {
      "get": {
        "tags": ["Thumbnails"],
        "summary": "Get thumbnail URL by key",
        "description": "Returns a signed download URL for a thumbnail by its B2 key. Auth is optional for `shares/` prefixed keys.",
        "parameters": [
          { "name": "key", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Signed thumbnail URL", "content": { "application/json": { "schema": { "type": "object", "properties": { "url": { "type": "string", "format": "uri" } } } } } },
          "403": { "description": "Forbidden — key doesn't belong to user" }
        }
      }
    },
    "/api/objects/thumbnail/batch": {
      "post": {
        "tags": ["Thumbnails"],
        "summary": "Batch-fetch thumbnail content",
        "description": "Returns signed thumbnail URLs for multiple objects at once.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "keys": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Batch thumbnail URLs" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "adminCookieAuth": {
        "type": "apiKey", "in": "cookie", "name": "__Host-Xenode_admin_session",
        "description": "Drive Admin JWT cookie, validated against the current Admin record and session version. Development uses Xenode_admin_session. This is separate from user authentication."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key authentication. Generate keys via `POST /api/keys`."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Session cookie managed by better-auth. 7-day expiry, auto-refreshes every 24h."
      },
      "cronAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Cron job secret token (CRON_SECRET env var)."
      }
    },
    "schemas": {
...
      "StorageObject": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "key": { "type": "string", "description": "Opaque B2 object key" },
          "bucketId": { "type": "string" },
          "userId": { "type": "string" },
          "size": { "type": "integer" },
          "contentType": { "type": "string" },
          "encryptedContentType": { "type": "string" },
          "isFolder": { "type": "boolean" },
          "isEncrypted": { "type": "boolean" },
          "encryptedDEK": { "type": "string" },
          "iv": { "type": "string" },
          "encryptedName": { "type": "string" },
          "encryptedDisplayName": { "type": "string" },
          "isChunked": { "type": "boolean" },
          "tags": { "type": "array", "items": { "type": "string" } },
          "position": { "type": "integer" },
          "mediaCategory": { "type": "string" },
          "thumbnail": { "type": "string", "description": "B2 key for thumbnail" },
          "thumbnailUrl": { "type": "string", "format": "uri", "description": "Pre-signed thumbnail URL" },
          "optimizedKey": { "type": "string" },
          "optimizedUrl": { "type": "string", "format": "uri" },
          "optimizedEncryptedDEK": { "type": "string" },
          "optimizedIV": { "type": "string" },
          "optimizedSize": { "type": "integer" },
          "aspectRatio": { "type": "number" },
          "b2FileId": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}